+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   ([#7545](https://github.com/mitmproxy/mitmproxy/pull/7545), @mhils)
 - Add a log message to point Docker mitmweb users towards `web_password`.
   ([#7554](https://github.com/mitmproxy/mitmproxy/pull/7554), @mhils)
+- Fix a bug where UTF-8 surrogates would crash the export addon.
 
 ## 06 February 2025: mitmproxy 11.1.2